more built out

+++ db/migrate/20200826035802_create_barbers.rb
@@ -2,8 +2,8 @@ class CreateBarbers < ActiveRecord::Migration[6.0]
   def change
     create_table :barbers do |t|
       t.string :name
-      t.integer :longlitude
-      t.integer :latitude
+      t.float :longlitude
+      t.float :latitude
 
       t.timestamps
     end
